requirements.txt
imageio==2.9.0
pillow >= 8.0